pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["cauesmelo <cauesmelo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic-settings = "2.1.0"
asyncpg = "0.29.0"
tortoise-orm = "^0.21.7"
aerich = "0.7.2"
httpx = "0.26.0"
pytest = "7.4.4"
gunicorn = "^23.0.0"
pytest-cov = "^6.0.0"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
pytest-xdist = "^3.6.1"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.3.1"

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
